Update BossClick.as
I updated how BossClick attempts to avoid pop-up blockers.  Instead of looking for various browsers by name, BossClick now looks for the Flash "player type" to see if it is ActiveX or Mac.  If the "player type" is ActiveX, BossClick opens the clickThrough link in a window by making a Javascript call from the browser instead of using the inherit Flash method of opening a browser window.